Add a deletion strategy for sequel

+++ lib/database_cleaner/sequel/truncation.rb
@@ -1,4 +1,4 @@
-require "database_cleaner/generic/truncation"
+require 'database_cleaner/generic/truncation'
 require 'database_cleaner/sequel/base'
 
 module DatabaseCleaner
@@ -7,15 +7,21 @@ module DatabaseCleaner
       include ::DatabaseCleaner::Sequel::Base
       include ::DatabaseCleaner::Generic::Truncation
 
+      def start
+        @last_txid = txid
+      end
+
       def clean
+        return unless dirty?
+
         case db.database_type
         when :postgres
           # PostgreSQL requires all tables with FKs to be truncates in the same command, or have the CASCADE keyword
           # appended. Bulk truncation without CASCADE is:
           # * Safer. Tables outside of tables_to_truncate won't be affected.
           # * Faster. Less roundtrips to the db.
-          unless (tables= tables_to_truncate(db)).empty?
-            all_tables= tables.map{|t| %["#{t}"]}.join ','
+          unless (tables = tables_to_truncate(db)).empty?
+            all_tables = tables.map { |t| %("#{t}") }.join(',')
             db.run "TRUNCATE TABLE #{all_tables};"
           end
         else
@@ -44,13 +50,24 @@ module DatabaseCleaner
         end
       end
 
+      def dirty?
+        @last_txid != txid || @last_txid.nil?
+      end
+
+      def txid
+        case db.database_type
+        when :postgres
+          db.fetch('SELECT txid_snapshot_xmax(txid_current_snapshot()) AS txid').first[:txid]
+        end
+      end
+
       def tables_to_truncate(db)
         (@only || db.tables.map(&:to_s)) - @tables_to_exclude
       end
 
       # overwritten
       def migration_storage_names
-        %w[schema_info schema_migrations]
+        %w(schema_info schema_migrations)
       end
 
       def pre_count?
